+++ src/generate.py
@@ -5,9 +5,9 @@ import yaml
 
 from datetime import datetime
 from transformers import PreTrainedTokenizerFast
-from architecture.lightning.performer import PerformerLightning
-from architecture.lightning.retnet import RetNetLightning
-from architecture.lightning.transformer import TransformerLightning
+from architecture.lightning.performer_lightning import PerformerLightning
+from architecture.lightning.retnet_lightning import RetNetLightning
+from architecture.lightning.transformer_lightning import TransformerLightning
 from utils import Struct, generate_text
 
 def generate_specific_text(config: Struct):
