+++ src/inference.py
@@ -9,9 +9,9 @@ from pathlib import Path
 from pytorch_lightning import Trainer, loggers as pl_loggers
 from pytorch_lightning.plugins.environments import SLURMEnvironment
 from torchinfo import summary as model_summary
-from architecture.lightning.performer import PerformerLightning
-from architecture.lightning.retnet import RetNetLightning
-from architecture.lightning.transformer import TransformerLightning
+from architecture.lightning.performer_lightning import PerformerLightning
+from architecture.lightning.retnet_lightning import RetNetLightning
+from architecture.lightning.transformer_lightning import TransformerLightning
 from utils import Struct
